be/requirements.txt
django
djangorestframework
black
virtualenv
djangorestframework-simplejwt
psycopg2
django-environ